+++ src/host/run.cpp
@@ -618,7 +618,8 @@ namespace ccf
       config.snapshots.directory,
       writer_factory,
       config.snapshots.read_only_directory,
-      config.snapshots.max_retained_snapshot_files);
+      config.snapshots.max_retained_snapshot_files,
+      config.snapshots.snapshot_cleanup_interval);
     snapshots.register_message_handlers(buffer_processor.get_dispatcher());
 
     // handle LFS-related messages from the enclave